requirements.txt
pynput
pywin32
python-dotenv
